scholarship/requirements.txt
asgiref==3.7.2
channels==4.0.0
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
easy==0.0.4
filelock==3.13.3
Jinja2==2.7.3
MarkupSafe==2.1.5
mysqlclient==2.2.1
Pillow==10.1.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==3.11
setuptools==69.0.3
sh==2.0.6
sqlparse==0.4.4
tzdata==2023.4
virtualenv==20.25.1
